+++ snips_nlu/default_configs/config_ja.py
@@ -85,7 +85,21 @@ CONFIG = {
                     {
                         "args": {
                             "use_stemming": False,
-                            "tagging_scheme_code": 2
+                            "tagging_scheme_code": 2,
+                            "entity_filter": {
+                                "automatically_extensible": False
+                            }
+                        },
+                        "factory_name": "entity_match",
+                        "offsets": [-1, 0, 1, 2],
+                    },
+                    {
+                        "args": {
+                            "use_stemming": False,
+                            "tagging_scheme_code": 2,
+                            "entity_filter": {
+                                "automatically_extensible": False
+                            }
                         },
                         "factory_name": "entity_match",
                         "offsets": [-1, 0, 1, 2],
@@ -144,7 +158,8 @@ CONFIG = {
                         "unknown_words_replacement_string": None,
                         "keep_order": True
                     }
-                }
+                },
+                "noise_reweight_factor": 1,
             }
         }
     ]