+++ snips_nlu/tests/test_nlu_engine.py
@@ -3,13 +3,16 @@ from __future__ import unicode_literals
 
 import io
 import shutil
+import sys
 from builtins import str
+from unittest import skipIf
 
+from checksumdir import dirhash
 from mock import MagicMock, patch
 from snips_nlu_parsers import get_all_languages
 
 import snips_nlu
-from snips_nlu import load_resources
+from snips_nlu.common.io_utils import temp_dir
 from snips_nlu.constants import (
     END, LANGUAGE, LANGUAGE_EN, RES_ENTITY, RES_INPUT, RES_INTENT,
     RES_INTENT_NAME, RES_MATCH_RANGE, RES_RAW_VALUE, RES_SLOTS, RES_SLOT_NAME,
@@ -24,6 +27,7 @@ from snips_nlu.intent_parser import IntentParser
 from snips_nlu.nlu_engine import SnipsNLUEngine
 from snips_nlu.pipeline.configs import (
     NLUEngineConfig)
+from snips_nlu.resources import load_resources
 from snips_nlu.result import (
     custom_slot, empty_result, intent_classification_result, parsing_result,
     resolved_slot, unresolved_slot, extraction_result)
@@ -1350,3 +1354,44 @@ utterances:
 
         # Then
         mocked_build_parser.assert_not_called()
+
+    @skipIf(sys.version_info[0:2] < (3, 5),
+            "The bug fixed here "
+            "https://github.com/scikit-learn/scikit-learn/pull/13422 is "
+            "available for scikit-learn>=0.21.0 in which the support for "
+            "Python<=3.4 has been dropped")
+    def test_training_should_be_reproducible(self):
+        # Given
+        random_state = 42
+        dataset_stream = io.StringIO("""
+---
+type: intent
+name: MakeTea
+utterances:
+- make me a hot cup of tea
+- make me five tea cups
+
+---
+type: intent
+name: MakeCoffee
+utterances:
+- make me one cup of coffee please
+- brew two cups of coffee""")
+        dataset = Dataset.from_yaml_files("en", [dataset_stream]).json
+
+        # When
+        engine1 = SnipsNLUEngine(random_state=random_state)
+        engine1.fit(dataset)
+
+        engine2 = SnipsNLUEngine(random_state=random_state)
+        engine2.fit(dataset)
+
+        # Then
+        with temp_dir() as tmp_dir:
+            dir_engine1 = tmp_dir / "engine1"
+            dir_engine2 = tmp_dir / "engine2"
+            engine1.persist(dir_engine1)
+            engine2.persist(dir_engine2)
+            hash1 = dirhash(str(dir_engine1), 'sha256')
+            hash2 = dirhash(str(dir_engine2), 'sha256')
+            self.assertEqual(hash1, hash2)